feat: replace filters file
