+++ frontend/views/containers/dashboard/GroupMembersAllModal.vue
@@ -121,7 +121,7 @@ import Search from '@components/Search.vue'
 import AvatarUser from '@components/AvatarUser.vue'
 import ProfileCard from '@components/ProfileCard.vue'
 import GroupMembersTooltipPending from '@containers/dashboard/GroupMembersTooltipPending.vue'
-import { filterByKeyword } from '@utils/filters.js'
+import { filterByKeyword } from '@view-utils/filters.js'
 
 export default ({
   name: 'GroupMembersAllModal',
